DSP-11 move map utils to own file
